Use qint64 instead of size_t for the folder size calculation, otherwise folders >2GB won't show up correctly on 32 Bit systems.
I hope I found all places were the conversion was necessary.
I couldn't really test this, as i don't have a 64 Bit system, but I should
work.

BUG: 151155


svn path=/trunk/KDE/kdepim/libkdepim/; revision=751020

+++ libkdepim-copy/kfoldertree.cpp
@@ -209,7 +209,7 @@ void KFolderTreeItem::setTotalCount( int aTotal )
 }
 
 //-----------------------------------------------------------------------------
-void KFolderTreeItem::setFolderSize( int aSize )
+void KFolderTreeItem::setFolderSize( qint64 aSize )
 {
   if ( aSize < 0 ) {
     return;  // we need to update even if nothing changed, kids ...
@@ -218,7 +218,7 @@ void KFolderTreeItem::setFolderSize( int aSize )
   mSize = aSize;
 
   QString size;
-  int recursiveSize = 0;
+  qint64 recursiveSize = 0;
 
   if ( childCount() > 0 && !isOpen() ) {
     recursiveSize = recursiveFolderSize();
@@ -246,9 +246,9 @@ void KFolderTreeItem::setFolderSize( int aSize )
 }
 
 //-----------------------------------------------------------------------------
-size_t KFolderTreeItem::recursiveFolderSize() const
+qint64 KFolderTreeItem::recursiveFolderSize() const
 {
-  int size = mSize;
+  qint64 size = mSize;
 
   for ( Q3ListViewItem *item = firstChild();  item; item = item->nextSibling() ) {
     size += static_cast<KFolderTreeItem*>( item )->recursiveFolderSize();
